chore: move HttpHeader to own file
