+++ packages/core/src/httpClient/httpHeaders.ts
@@ -1,0 +1,3 @@
+export interface HttpHeaders {
+  [key: string]: string;
+}
